pyproject.toml
[project]
name = "tockar-discord-bot"
version = "0.1.0"
description = "A Discord bot built with discord.py"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.4.0",
    "dotenv>=0.9.9",
    "environ-config>=24.0.0",
    "wheelspin-gif>=1.0.0",
    "numpy>=1.21.0",
    "aiohttp>=3.8.0",
]
